feat!: test GetTableSchema via regular select queries (#45)
## What's Changed

Test GetTableSchema by using existing test cases for `type/select` which
are almost the same. The main difference is not having a fixed table
name and having an extra sort column, which are worked around in the
test case itself.

This requires a bit more trickery on our side but saves dev effort in
not having to effectively duplicate a lot of test cases.

Fixes #43.