docs: ensure footnotes work in raw HTML types table (#115)
## What's Changed

This is the necessary fix to make footnotes work correctly in our
generated validation pages. Before this, footnotes didn't work when the
footnote reference was inside a raw HTML table without special handling.
The special handling is to add blank lines around the templates and
ensure the cell content isn't indented any more than myst-parser
tolerates. Apparently this is 0-3 spaces. Once you indent it to 4
spaces, footnotes break.

For this change, we just don't indent and surround with blank lines.

Closes #114